pyproject.toml
[tool.poetry]
name = "trips-api"
version = "0.1.0"
description = "Microservicio de viajes"
authors = ["Daniel Lovera <daanilo2497@gmail.com>"]
readme = "README.md"
license = "MIT"

packages = [
    { include = "script", from = "src" }
]

[tool.pytest.ini_options]
pythonpath = [
  ".", "src",
]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.82.0"
uvicorn = {extras = ["standard"], version = "^0.18.3"}
psycopg2-binary = "^2.9.4"


[tool.poetry.group.dev.dependencies]
mypy = "^0.971"
pytest = "^7.1.3"
requests = "^2.28.1"
black = "^22.8.0"

[tool.poetry.scripts]
format = "script.poetry:format"
reformat = "script.poetry:reformat"
lint = "script.poetry:lint"
start = "script.poetry:start"
test = "script.poetry:test"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
